Organized multiple files into the helpers folder
