+++ src/groupsIndexParser.js
